Refactor register & implement create and edit action for task (domain, repository, service & controller)

+++ agendaCraft/src/main/java/com/agendaCraft/agendaCraft/repository/TaskRepository.java
@@ -1,12 +1,10 @@
 package com.agendaCraft.agendaCraft.repository;
 
+
 import com.agendaCraft.agendaCraft.domain.Task;
-import com.agendaCraft.agendaCraft.domain.User;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-
 @Repository
-public interface TaskRepository extends JpaRepository<Task, Integer> {
+public interface TaskRepository extends JpaRepository<Task, Long> {
 }
